% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_choropleth.R
\name{lt_choropleth}
\alias{lt_choropleth}
\title{Leaflet choropleths}
\usage{
lt_choropleth(data = NULL, map_name = NULL, var = NULL, filter = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
ggplot viz
}
\description{
Leaflet choropleths
}
\section{ctypes}{
 Gcd-Num
}

\examples{
gg_choropleth_GcdNum(sample_data("Gcd-Num", nrow = 10))
}
