% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltgeo-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{available_maps}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{geodato}{\code{\link[geodato]{available_maps}}}
}}

