% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-utils.R
\docType{package}
\name{ltgeo-package}
\alias{ltgeo}
\alias{ltgeo-package}
\title{ltgeo: Leaflet Wrapper To Make Interactive Maps}
\description{
Make interactive maps with leaflet with a simplified interface.
}
\author{
\strong{Maintainer}: Jp Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
